Corrige errores de nomenclatura y mejora la estructura de paquetes en las clases de respuesta y solicitud

+++ src/main/java/co/edu/uniquindio/oldbaker/dto/ProductoResponse.java
@@ -14,6 +14,6 @@ public class ProductoResponse {
     // Receta asociada
     private Long idReceta;
     private String insumoNombre;
-    private Integer cantidadInsumo;
+    private Double cantidadInsumo;
 }
